Split into smaller file - separation of concern
